# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Osau AMS.",
        "title": "Osau AMS API",
        "contact": {
            "name": "API Support",
            "email": "bakit.german.work@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/specialties": {
            "get": {
                "description": "Get a list of all specialties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Get all specialties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.SpecialtyInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing specialty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Update a specialty",
                "parameters": [
                    {
                        "description": "Specialty info",
                        "name": "specialty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PutSpecialtyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new specialty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Create a specialty",
                "parameters": [
                    {
                        "description": "Specialty info",
                        "name": "specialty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSpecialtyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially update an existing specialty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Partially update a specialty",
                "parameters": [
                    {
                        "description": "Specialty info",
                        "name": "specialty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PatchSpecialtyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/specialties/departament/{departament_id}": {
            "get": {
                "description": "Get a list of specialties by departament ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Get specialties by departament ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Departament ID",
                        "name": "departament_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.SpecialtyInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/specialties/name/{name}": {
            "get": {
                "description": "Get a specialty by its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Get a specialty by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specialty Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SpecialtyInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/specialties/{code}": {
            "get": {
                "description": "Get a specialty by its code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Get a specialty by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specialty Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SpecialtyInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing specialty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Specialties"
                ],
                "summary": "Delete a specialty",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Specialty Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Specialty": {
            "type": "object",
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "education_level_id": {
                    "type": "integer"
                },
                "specialty_code": {
                    "type": "string"
                },
                "specialty_name": {
                    "type": "string"
                }
            }
        },
        "domain.SpecialtyInfo": {
            "type": "object",
            "properties": {
                "specialty": {
                    "$ref": "#/definitions/domain.Specialty"
                },
                "specialty_sub": {
                    "$ref": "#/definitions/domain.SpecialtySub"
                }
            }
        },
        "domain.SpecialtySub": {
            "type": "object",
            "properties": {
                "departament_name": {
                    "type": "string"
                },
                "education_level_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.CreateSpecialtyRequest": {
            "type": "object",
            "required": [
                "departament_id",
                "education_level_id",
                "specialty_code",
                "specialty_name"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "education_level_id": {
                    "type": "integer"
                },
                "specialty_code": {
                    "type": "string"
                },
                "specialty_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PatchSpecialtyRequest": {
            "type": "object",
            "required": [
                "education_level_id",
                "specialty_code"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "education_level_id": {
                    "type": "integer"
                },
                "specialty_code": {
                    "type": "string"
                },
                "specialty_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.PutSpecialtyRequest": {
            "type": "object",
            "required": [
                "departament_id",
                "education_level_id",
                "specialty_code",
                "specialty_name"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "education_level_id": {
                    "type": "integer"
                },
                "specialty_code": {
                    "type": "string"
                },
                "specialty_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}